Refactored the state-full `FormatSelector` into `BcoinUtils` which is a folder of independent files, each contains only stateless utility functions. Also moved the `paymentRequest` into `BcoinUtils`.
